eliminate SingleParseResult (always use ParseResult)
ParseResult now has a node and a sequence of remaining tokens.
A ParseResult can easily be extended by applying additional
symbol application functions (and adding the resulting
child nodes to the overall node in progress.)

This change eliminates the distinction between do-production
and apply-production, and allows do-production to be used
in cases where a partial production (e.g., a common prefix
on the right hand side of multiple productions) is desired.

apply-production is now called continue-production, which
better reflects its purpose.