support optional parser context
This is useful for allowing parse functions to have
different behavior depending on the context in which
they are used.  For example, we want parse-declaration
to allow declarations at file scope to be function
definitions (for the first declarator only), but we
won't want function definitions to be allowed within
a compound statement.