support parsing function definitions
parse-init-declarator (which I suppose should really
be called parse-init-declarator-or-function-definition)
checks to see whether a compound statement is present
after a declarator (assuming that it is the first
overall declarator and is in a context where a
function definition is allowed), and if so, parses
it as a function definition.

Amazingly this worked on the first try!