renamed cljs files to cljc
Goal is to support both Clojure and ClojureScript.

Right now, the only place this makes a difference is
catparty.exc (throwing exceptions).  exc.cljs is left
as ClojureScript.  (Eventually there will be an equivalent
for Clojure.)

Unfortunately it doesn't seem to be possible to have
multiple source directories in a ClojureScript project.
(Would be nice to have src-cljc, src-cljs, src-clj, etc.)