initial support for parsing abstract declarators
The idea is
- to use the context to indicate explicitly whether an
  abstract or concrete declarator is allowed (or either)
- to detect in parse-direct-declarator-base the point
  at which an epsilon production can be applied to mark
  the point equivalent to an identifier in a concrete
  declarator

The main issue is that a left paren could indicate
either grouping (with a nested declarator) or a
suffix (indicating a function).  The way to resolve this
ambiguity is to note that in the case of the function
suffix, the token after the left paren will be a token
that can start a type name.