fix bugs in param decl parsing, declarator grouping
The code to detect when a declarator in a parameter declaration
was omitted was incorrectly treating a '(' lookahead after
the declaration specifiers as indicating a missing declarator.
(It was supposed to be ')'.)

The code to use parentheses for grouping/precedence in
direct declarator base parsing had omitted the code to consume
the parentheses.

After these changes, parentheses in a direct declarator base
seem to be parsed correctly.